+++ test/utils/utils.dart
@@ -1,4 +1,5 @@
 export 'utils_logs.dart';
 export 'utils_main.dart';
+export 'utils_style.dart';
 export 'utils_format.dart';
 export 'utils_network.dart';