+++ test/utils/utils_main.dart
@@ -3,6 +3,7 @@ import '../test.dart';
 class UtilsTestMain {
   static void main() {
     utilsLogs();
+    utilsStyle();
     utilsFormat();
     utilsNetwork();
   }