+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.0.6] - GraphQl support & Floating Style
+* [update] Add floating style 
+* [fix] Update cURL copy support GraphQL request
+
 ## [0.0.5] - Documentation
 * [update] Example documentation main
 * [fix] Assets image read.me